requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.9.1+cpu
sentence-transformers
transformers
accelerate
numpy
fastapi
uvicorn
mysql-connector-python
requests
beautifulsoup4
lxml
python-dotenv
